Remove console.log. Move flashcards hook into flashcards component.
